preference name changed to TaskDescriptionBarColor to avoid conflict with same preferences used in cordova-plugin-statusbar
Preference name changed to TaskDescriptionBarColor to avoid conflict with same preferences used in cordova-plugin-statusbar.  The status bar color set via StatusBarBackgroundColor in the cordoba-plugin-status plugin sets the top status bar color which is the dark primary color according to the material design spec.  The task description bar color is normally the primary color. Hence the identical preference name used in this plugin was in conflict.